requirements.txt
bcrypt==4.2.0
cachetools==5.5.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
numpy==2.1.2
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
python-dotenv==1.0.1
Pyrebase4==4.8.0
requests==2.32.3
requests-oauthlib==2.0.0
streamlit==1.39.0
streamlit-cookies-manager==0.2.0
six==1.16.0
urllib3
